requirements.txt
cryptography==3.4.8
fastapi==0.68.2
pydantic==1.9.0
sqlalchemy==1.4.31
uvicorn==0.15.0
python-dotenv==0.19.1
jose==1.0.0
